chore: trim verbose comments; extract silence-review and share-reaction logic
Trim AI-style comments across the desktop editor panels, overlays, settings,
and (app) routes: drop code-restating notes, narration, and decorative
banners; condense multi-sentence blocks to one or two lines. Parity notes,
platform gotchas, lint directives, and exported JSDoc kept. Comment-only there.

Also pull two testable units out of their components: the silence-review
detection presets, confidence classification, and cut-bounds maths into
silence-review.logic.ts, and the optimistic share-page reaction toggle into
share/engagement.ts (both unit-tested).
