+++ apps/desktop/src/components/editor/properity-panel/info-panel.logic.ts
@@ -1,15 +1,11 @@
-/**
- * Pure helpers for InfoPanel: relative-time labels, path basename, and
- * annotation-kind counts. Extracted so they're testable; the component keeps the
- * store reads (resolution/fps/duration) and the icon table.
- */
+/** InfoPanel helpers: relative-time labels, path basename, and annotation-kind counts. */
 
 import type { Annotation } from "$lib/stores/editor-store.svelte";
 
 /**
- * Chat-style relative time vs a `current` epoch (ms): "just now", "5s",
- * "3 min", "2 hr", "4 days", prefixed "in "/suffixed " ago" by direction.
- * Floors at each cutoff so the readout doesn't bounce by a unit each tick.
+ * Chat-style relative time vs a `current` epoch (ms): "just now", "5s", "3 min",
+ * "2 hr", "4 days", prefixed "in "/suffixed " ago". Floors at each cutoff so the
+ * readout doesn't bounce by a unit each tick.
  */
 export function formatRelative(ts: number | null, current: number): string {
 	if (!ts) return "Never";
@@ -37,10 +33,7 @@ export function basename(path: string): string {
 	return last;
 }
 
-/**
- * Count annotations by kind, always seeding every kind to 0 so the readout row
- * doesn't shift as shapes are added/removed.
- */
+/** Count annotations by kind; seeds every kind to 0 so the readout row doesn't shift. */
 export function countByKind(annotations: Annotation[]): Record<string, number> {
 	const out: Record<string, number> = {
 		rect: 0,
